refactor: split test_mvbart.py
