+++ tests/test_mcmcstep.py
@@ -37,10 +37,23 @@ from jax import debug_key_reuse, make_mesh, random, tree, vmap
 from jax import numpy as jnp
 from jax.sharding import AxisType, Mesh, PartitionSpec, SingleDeviceSharding
 from jax.tree_util import KeyPath, keystr
-from jaxtyping import Array, Bool, Int32, Key, PyTree, UInt8, jaxtyped
-from numpy.testing import assert_array_equal
+from jaxtyping import (
+    Array,
+    Bool,
+    Float,
+    Float32,
+    Int32,
+    Key,
+    PyTree,
+    UInt8,
+    UInt32,
+    jaxtyped,
+)
+from numpy.testing import assert_allclose, assert_array_equal
+from pytest import FixtureRequest  # noqa: PT013
 from pytest_subtests import SubTests
 from scipy import stats
+from scipy.stats import chi2, ks_1samp, ks_2samp
 
 from bartz.jaxext import get_device_count, minimal_unsigned_dtype, split
 from bartz.mcmcstep import State, init, step
@@ -51,7 +64,21 @@ from bartz.mcmcstep._moves import (
     split_range,
 )
 from bartz.mcmcstep._state import chain_vmap_axes, data_vmap_axes
-from bartz.mcmcstep._step import step_error_cov_inv, step_trees, step_z
+from bartz.mcmcstep._step import (
+    Counts,
+    _compute_likelihood_ratio_mv,
+    _compute_likelihood_ratio_uv,
+    _precompute_leaf_terms_mv,
+    _precompute_leaf_terms_uv,
+    _precompute_likelihood_terms_mv,
+    _precompute_likelihood_terms_uv,
+    _sample_wishart_bartlett,
+    _step_error_cov_inv_mv,
+    _step_error_cov_inv_uv,
+    step_error_cov_inv,
+    step_trees,
+    step_z,
+)
 from tests.util import assert_close_matrices, manual_tree
 
 
@@ -1083,3 +1110,453 @@ class TestMixedBinaryContinuous:
         init_kwargs.update(outcome_type=['binary'] * (self.k - 1))
         with pytest.raises(AssertionError):
             init(**init_kwargs)
+
+
+class MCMCStepData(NamedTuple):
+    """Toy dataset for testing."""
+
+    X: Int32[Array, 'p n']
+    y: Float32[Array, ' n']
+    max_split: UInt32[Array, ' p']
+
+
+def random_pd_matrix(key: Key[Array, ''], k: int) -> Float[Array, '{k} {k}']:
+    """Generate a random positive definite matrix."""
+    A = random.normal(key, (k, k))
+    return A @ A.T + jnp.eye(k)
+
+
+@pytest.fixture(params=[(10, 2), (20, 5), (3, 100), (50, 50)])
+def mcmcstep_data_shape(request: FixtureRequest) -> tuple[int, int]:
+    """Provide (n, p) pairs for testing."""
+    return request.param
+
+
+@pytest.fixture
+def mcmcstep_data(mcmcstep_data_shape: tuple[int, int]) -> MCMCStepData:
+    """Generate a toy dataset."""
+    n, p = mcmcstep_data_shape
+    X = jnp.arange(n * p).reshape(p, n)
+    y = jnp.linspace(-1, 1, n)
+    max_split = jnp.full(p, 5, dtype=jnp.uint32)
+    return MCMCStepData(X, y, max_split)
+
+
+class TestWishart:
+    """Test the basic properties of the wishart sampler output."""
+
+    # Parameterize with (k, df) pairs
+    @pytest.fixture(params=[(1, 3), (3, 3), (3, 5), (3, 100), (100, 102)])
+    def wishart_params(self, request: FixtureRequest) -> tuple[int, int]:
+        """Provide (k, df) pairs for testing."""
+        k, df = request.param
+        return k, df
+
+    def ill_conditioned_matrix(
+        self, key: Key[Array, ''], k: int, condition_number: float = 1e6
+    ) -> Float[Array, '{k} {k}']:
+        """Generate a ill conditioned random positive semi-definite matrix."""
+        A = random.normal(key, (k, k))
+        U, _ = jnp.linalg.qr(A)
+
+        if k == 1:
+            eigs = jnp.zeros(1)
+        else:
+            smalls = jnp.geomspace(1.0, 1.0 / condition_number, num=k - 1)
+            eigs = jnp.concatenate([smalls, jnp.array([0.0])])
+        return (U * eigs) @ U.T
+
+    def test_size(self, keys: split, wishart_params: tuple[int, int]) -> None:
+        """Check that the sample generated by wishart sampler is of shape k*k."""
+        k, df = wishart_params
+        scale = random_pd_matrix(keys.pop(), k)
+        sample = _sample_wishart_bartlett(keys.pop(), df, scale)
+        assert sample.shape == (k, k)
+
+    def test_symmetric(self, keys: split, wishart_params: tuple[int, int]) -> None:
+        """Check that the sample generated by wishart sampler is symmetric."""
+        k, df = wishart_params
+        scale = random_pd_matrix(keys.pop(), k)
+        sample = _sample_wishart_bartlett(keys.pop(), df, scale)
+        assert_close_matrices(sample, sample.T, rtol=1e-6)
+
+    def test_pos_def(self, keys: split, wishart_params: tuple[int, int]) -> None:
+        """Check that the sample generated by wishart sampler is positive definite."""
+        k, df = wishart_params
+        scale = random_pd_matrix(keys.pop(), k)
+        sample = _sample_wishart_bartlett(keys.pop(), df, scale)
+        eigs = jnp.linalg.eigvalsh(sample)
+        assert jnp.all(eigs > 0)
+
+    def test_near_singular_scale(
+        self, keys: split, wishart_params: tuple[int, int]
+    ) -> None:
+        """Check that the wishart sampler still works with singular or near singular matrix."""
+        k, df = wishart_params
+        ill_conditioned_scale = self.ill_conditioned_matrix(keys.pop(), k)
+        sample = _sample_wishart_bartlett(keys.pop(), df, ill_conditioned_scale)
+        assert jnp.all(jnp.isfinite(sample))
+
+    def test_wishart_dist(self, keys: split, wishart_params: tuple[int, int]) -> None:
+        """Check that the sample generated by wishart sampler follows a wishart distribution."""
+        k, df = wishart_params
+        sigma = random_pd_matrix(keys.pop(), k)
+        scale_inv = jnp.linalg.inv(sigma)
+
+        a = random.normal(keys.pop(), (k,))
+        denominator = a.T @ sigma @ a
+
+        sampler = vmap(_sample_wishart_bartlett, in_axes=(0, None, None))
+        W = sampler(keys.pop(1000), float(df), scale_inv)
+        t = jnp.einsum('ijk,j,k->i', W, a, a) / denominator
+
+        test = ks_1samp(t, chi2(df).cdf)
+        assert test.pvalue > 0.01
+
+
+class TestPrecomputeTerms:
+    """Test _precompute_likelihood_terms_mv and _precompute_leaf_terms_mv correctness and stability."""
+
+    @pytest.fixture(params=[1, 2, 5, 10])
+    def k(self, request: FixtureRequest) -> int:
+        """Provide different ks for testing."""
+        return request.param
+
+    def test_shapes_leaf(self, keys: split, k: int) -> None:
+        """Check that shapes of outputs are correct."""
+        num_trees, tree_size = 3, 4
+        prec_trees = jnp.ones((num_trees, tree_size))
+        error_cov_inv = random_pd_matrix(keys.pop(), k)
+        leaf_prior_cov_inv = random_pd_matrix(keys.pop(), k)
+
+        result = _precompute_leaf_terms_mv(
+            keys.pop(), prec_trees, error_cov_inv, leaf_prior_cov_inv
+        )
+        assert result.mean_factor.shape == (num_trees, k, k, tree_size)
+        assert result.centered_leaves.shape == (num_trees, k, tree_size)
+
+    def test_likelihood_equiv(self, keys: split) -> None:
+        """Check that _compute_likelihood_ratio_uv and _compute_likelihood_ratio_mv agree when k = 1."""
+        inv_sigma2 = random.uniform(keys.pop(), (), minval=0.1, maxval=5.0)
+        leaf_prior_cov_inv_uv = random.uniform(keys.pop(), (), minval=0.1, maxval=5.0)
+        error_cov_inv = jnp.array([[inv_sigma2]])
+        leaf_prior_cov_inv = jnp.array([[leaf_prior_cov_inv_uv]])
+
+        precs = Counts(left=jnp.array(3.0), right=jnp.array(4.0), total=jnp.array(7.0))
+
+        total_resid = random.normal(keys.pop(), (1,))
+        left_resid = random.normal(keys.pop(), (1,))
+        right_resid = random.normal(keys.pop(), (1,))
+
+        prelkv_mv, _ = _precompute_likelihood_terms_mv(
+            error_cov_inv, leaf_prior_cov_inv, precs
+        )
+        likelihood_mv = _compute_likelihood_ratio_mv(
+            total_resid, left_resid, right_resid, prelkv_mv
+        )
+
+        prelkv_uv, prelk_uv = _precompute_likelihood_terms_uv(
+            inv_sigma2, leaf_prior_cov_inv_uv, precs
+        )
+        likelihood_uv = _compute_likelihood_ratio_uv(
+            total_resid.item(),
+            left_resid.item(),
+            right_resid.item(),
+            prelkv_uv,
+            prelk_uv,
+        )
+
+        assert_allclose(
+            prelkv_mv.log_sqrt_term, prelkv_uv.log_sqrt_term, rtol=1e-6, atol=1e-6
+        )
+        assert_allclose(likelihood_mv, likelihood_uv, rtol=1e-6, atol=1e-6)
+
+    def test_leaf_terms_equiv(self, keys: split) -> None:
+        """Check that _precompute_leaf_terms_uv and _precompute_leaf_terms_mv agree when k = 1."""
+        num_trees, tree_size = 2, 3
+        inv_sigma2 = random.uniform(keys.pop(), (), minval=0.1, maxval=5.0)
+        leaf_prior_cov_inv_uv = random.uniform(keys.pop(), (), minval=0.1, maxval=5.0)
+
+        error_cov_inv = jnp.array([[inv_sigma2]])
+        leaf_prior_cov_inv = jnp.array([[leaf_prior_cov_inv_uv]])
+        prec_trees = random.uniform(keys.pop(), (num_trees, tree_size)) * 5.0
+        z_mv = random.normal(keys.pop(), (num_trees, tree_size, 1))
+        z_uv = z_mv.squeeze(axis=-1)
+
+        result_uv = _precompute_leaf_terms_uv(
+            keys.pop(), prec_trees, inv_sigma2, leaf_prior_cov_inv_uv, z_uv
+        )
+        result_mv = _precompute_leaf_terms_mv(
+            keys.pop(), prec_trees, error_cov_inv, leaf_prior_cov_inv, z_mv
+        )
+
+        assert_allclose(
+            result_uv.mean_factor,
+            result_mv.mean_factor.squeeze((1, 2)),
+            rtol=1e-6,
+            atol=1e-6,
+        )
+        assert_allclose(
+            result_uv.centered_leaves,
+            result_mv.centered_leaves.squeeze(1),
+            rtol=1e-6,
+            atol=1e-6,
+        )
+
+
+class TestMVBartIntegration:
+    """Test equivalence between Univariate and Multivariate (k=1) modes."""
+
+    @pytest.mark.parametrize('binary', [False, True])
+    def test_init_equivalence(self, mcmcstep_data: MCMCStepData, binary: bool) -> None:
+        """Test that init produces compatible structures for UV and MV(k=1)."""
+        X, y, max_split = mcmcstep_data
+        p_nonterminal = jnp.array([0.9, 0.5])
+
+        if binary:
+            y = (y > 0).astype(jnp.float32)
+
+        common: dict = dict(
+            X=X,
+            max_split=max_split,
+            num_trees=10,
+            p_nonterminal=p_nonterminal,
+            resid_num_batches=None,
+            count_num_batches=None,
+        )
+
+        uv_kw = dict(y=y, offset=0.0, leaf_prior_cov_inv=1.0)
+        mv_kw = dict(
+            y=y[None, :], offset=jnp.zeros(1), leaf_prior_cov_inv=jnp.array([[1.0]])
+        )
+
+        if binary:
+            uv_kw.update(outcome_type='binary')
+            mv_kw.update(outcome_type='binary')
+        else:
+            uv_kw.update(error_cov_df=6.0, error_cov_scale=4.0)
+            mv_kw.update(error_cov_df=jnp.array(6.0), error_cov_scale=4.0 * jnp.eye(1))
+
+        bart_uv = init(**uv_kw, **common)
+        bart_mv = init(**mv_kw, **common)
+
+        assert bart_uv.resid.ndim == 1
+        assert bart_mv.resid.ndim == 2
+        assert bart_mv.resid.shape[0] == 1
+        assert bart_mv.resid.shape[1] == bart_uv.resid.shape[0]
+
+        assert jnp.ndim(bart_uv.error_cov_inv) == 0
+        assert bart_mv.error_cov_inv.shape == (1, 1)
+
+        if binary:
+            assert bart_uv.binary_y is not None
+            assert bart_mv.binary_y is not None
+            assert bart_uv.binary_y.ndim == 1
+            assert bart_mv.binary_y.ndim == 2
+            assert_array_equal(bart_uv.binary_y, bart_mv.binary_y.squeeze(0))
+            assert bart_uv.z is not None
+            assert bart_mv.z is not None
+            assert bart_uv.z.ndim == 1
+            assert bart_mv.z.ndim == 2
+            assert_array_equal(bart_uv.z, bart_mv.z.squeeze(0))
+
+        assert_array_equal(bart_uv.resid, bart_mv.resid.squeeze(0))
+        assert_array_equal(bart_uv.forest.var_tree, bart_mv.forest.var_tree)
+        assert_array_equal(bart_uv.forest.split_tree, bart_mv.forest.split_tree)
+        assert_array_equal(
+            bart_uv.forest.leaf_tree, bart_mv.forest.leaf_tree.squeeze(1)
+        )
+        assert_array_equal(bart_uv.forest.leaf_indices, bart_mv.forest.leaf_indices)
+        assert_array_equal(bart_uv.forest.p_nonterminal, bart_mv.forest.p_nonterminal)
+        assert_array_equal(bart_uv.forest.p_propose_grow, bart_mv.forest.p_propose_grow)
+        assert_array_equal(bart_uv.forest.affluence_tree, bart_mv.forest.affluence_tree)
+
+    def test_step_sigma_distribution_match(
+        self, keys: split, mcmcstep_data: MCMCStepData
+    ) -> None:
+        """
+        Test that _step_error_cov_inv_uv and _step_error_cov_inv_mv (k = 1) sample from the same posterior.
+
+        UV: 1/sigma2 ~ Gamma(alpha_post, beta_post)
+        MV: error_cov_inv ~ Wishart(df_post, scale_post)
+        """
+        X, y, _ = mcmcstep_data
+        resid = random.normal(keys.pop(), (y.size,))
+
+        # inverse gamma prior: alpha = df / 2, beta = scale / 2
+        df_prior = jnp.float32(20.0)
+        scale_prior = jnp.float32(10.0)
+
+        common: dict = dict(
+            X=X,
+            binary_y=None,
+            binary_indices=None,
+            error_cov_df=df_prior,
+            z=None,
+            offset=0.0,
+            prec_scale=None,
+            forest=None,
+            config=None,
+        )
+
+        st_uv = State(
+            **common, resid=resid, error_cov_scale=scale_prior, error_cov_inv=1.0
+        )
+
+        st_mv = State(
+            **common,
+            resid=resid[None, :],
+            error_cov_scale=jnp.array([[scale_prior]]),
+            error_cov_inv=jnp.eye(1),
+        )
+
+        def sample_uv(k: Key[Array, '']) -> Float32[Array, '']:
+            return _step_error_cov_inv_uv(k, st_uv).error_cov_inv
+
+        def sample_mv(k: Key[Array, '']) -> Float32[Array, '']:
+            return _step_error_cov_inv_mv(k, st_mv).error_cov_inv.reshape(())
+
+        n_samples = 10000
+        samples_uv = vmap(sample_uv)(keys.pop(n_samples))
+        samples_mv = vmap(sample_mv)(keys.pop(n_samples))
+
+        _, p_value = ks_2samp(samples_uv, samples_mv)
+
+        assert jnp.abs(jnp.mean(samples_uv) - jnp.mean(samples_mv)) < 0.01
+        assert p_value > 0.01
+
+
+class TestMVBartSteps:
+    """Test the full MCMC step trajectory (init + multiple steps)."""
+
+    @pytest.mark.parametrize('binary', [False, True])
+    def test_step_trees_exact_match(
+        self, keys: split, mcmcstep_data: MCMCStepData, binary: bool
+    ) -> None:
+        """Test that MV tree logic is Identical to UV logic."""
+        X, y, max_split = mcmcstep_data
+        n_trees = 100
+
+        if binary:
+            y = (y > 0).astype(jnp.float32)
+
+        params: dict = dict(
+            X=X,
+            max_split=max_split,
+            num_trees=n_trees,
+            p_nonterminal=jnp.array([0.9, 0.5]),
+            resid_num_batches=None,
+            count_num_batches=None,
+        )
+
+        uv_kw = dict(y=y, offset=0.0, leaf_prior_cov_inv=jnp.float32(n_trees))
+        mv_kw = dict(
+            y=y[None, :], offset=jnp.zeros(1), leaf_prior_cov_inv=n_trees * jnp.eye(1)
+        )
+
+        if binary:
+            uv_kw.update(outcome_type='binary')
+            mv_kw.update(outcome_type='binary')
+        else:
+            uv_kw.update(error_cov_df=4.0, error_cov_scale=2.0)
+            mv_kw.update(error_cov_df=jnp.array(4.0), error_cov_scale=2 * jnp.eye(1))
+
+        uv_state = init(**uv_kw, **params)
+        mv_state = init(**mv_kw, **params)
+
+        mv_state = replace(
+            mv_state,
+            resid=uv_state.resid[None, :],
+            error_cov_inv=jnp.array([[uv_state.error_cov_inv]]),
+            forest=replace(
+                mv_state.forest,
+                var_tree=uv_state.forest.var_tree,
+                split_tree=uv_state.forest.split_tree,
+                leaf_tree=uv_state.forest.leaf_tree[:, None, :],
+                leaf_indices=uv_state.forest.leaf_indices,
+                affluence_tree=uv_state.forest.affluence_tree,
+            ),
+        )
+
+        key = keys.pop()
+        uv_next = step_trees(key, uv_state)
+        mv_next = step_trees(random.clone(key), mv_state)
+
+        assert_close_matrices(
+            uv_next.resid, mv_next.resid.squeeze(0), atol=1e-6, rtol=1e-6
+        )
+        assert_close_matrices(
+            uv_state.forest.leaf_tree,
+            mv_state.forest.leaf_tree.squeeze(1),
+            atol=1e-6,
+            rtol=1e-6,
+        )
+
+        assert_array_equal(uv_state.forest.var_tree, mv_state.forest.var_tree)
+        assert_array_equal(uv_state.forest.split_tree, mv_state.forest.split_tree)
+        assert_array_equal(uv_state.forest.leaf_indices, mv_state.forest.leaf_indices)
+        assert_array_equal(
+            uv_state.forest.affluence_tree, mv_state.forest.affluence_tree
+        )
+
+        assert_array_equal(
+            uv_state.forest.grow_prop_count, mv_state.forest.grow_prop_count
+        )
+        assert_array_equal(
+            uv_state.forest.grow_acc_count, mv_state.forest.grow_acc_count
+        )
+        assert_array_equal(
+            uv_state.forest.prune_prop_count, mv_state.forest.prune_prop_count
+        )
+        assert_array_equal(
+            uv_state.forest.prune_acc_count, mv_state.forest.prune_acc_count
+        )
+
+    @pytest.mark.parametrize('binary', [False, True])
+    def test_mv_steps(
+        self, keys: split, mcmcstep_data: MCMCStepData, binary: bool
+    ) -> None:
+        """Test that mv mode can run without crashing."""
+        X, y_uv, max_split = mcmcstep_data
+        k = 3
+
+        if binary:
+            y = random.bernoulli(keys.pop(), 0.5, (k, y_uv.size)).astype(jnp.float32)
+        else:
+            y = jnp.tile(y_uv, (k, 1))
+            y = y + random.normal(keys.pop(), y.shape) * 0.1
+
+        kw: dict = dict(
+            X=X,
+            y=y,
+            offset=jnp.zeros(k),
+            max_split=max_split,
+            num_trees=5,
+            p_nonterminal=jnp.array([0.9, 0.5]),
+            leaf_prior_cov_inv=jnp.eye(k),
+            resid_num_batches=None,
+            count_num_batches=None,
+        )
+
+        if binary:
+            kw.update(outcome_type='binary')
+        else:
+            kw.update(error_cov_df=jnp.array(10.0), error_cov_scale=jnp.eye(k))
+
+        mv_state = init(**kw)
+
+        for key in keys.pop(10):
+            mv_state = step(key, mv_state)
+
+            assert jnp.all(jnp.isfinite(mv_state.resid))
+            assert jnp.all(jnp.isfinite(mv_state.forest.leaf_tree))
+
+            assert mv_state.resid.shape == (k, y.shape[1])
+
+            assert jnp.all(jnp.isfinite(mv_state.error_cov_inv))
+            assert mv_state.error_cov_inv.shape == (k, k)
+
+            if binary:
+                assert mv_state.z is not None
+                assert jnp.all(jnp.isfinite(mv_state.z))
+                assert mv_state.z.shape == (k, y.shape[1])
