--- tests/test_mvbart.py
@@ -1,892 +1,0 @@
-# bartz/tests/test_mvbart.py
-#
-# Copyright (c) 2025-2026, The Bartz Contributors
-#
-# This file is part of bartz.
-#
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-#
-# The above copyright notice and this permission notice shall be included in all
-# copies or substantial portions of the Software.
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-# SOFTWARE.
-
-"""Test multivariate BART components."""
-
-from dataclasses import replace
-from typing import NamedTuple
-
-import pytest
-from jax import debug_nans, random, vmap
-from jax import numpy as jnp
-from jaxtyping import Array, Float, Float32, Int32, Key, UInt32
-from numpy.testing import assert_allclose, assert_array_equal
-from pytest import FixtureRequest  # noqa: PT013
-from pytest_subtests import SubTests
-from scipy.stats import chi2, ks_1samp, ks_2samp
-
-from bartz import Bart
-from bartz.jaxext import split
-from bartz.mcmcstep import State, init, step
-from bartz.mcmcstep._step import (
-    Counts,
-    _compute_likelihood_ratio_mv,
-    _compute_likelihood_ratio_uv,
-    _precompute_leaf_terms_mv,
-    _precompute_leaf_terms_uv,
-    _precompute_likelihood_terms_mv,
-    _precompute_likelihood_terms_uv,
-    _sample_wishart_bartlett,
-    _step_error_cov_inv_mv,
-    _step_error_cov_inv_uv,
-    step_trees,
-)
-from bartz.testing import gen_data
-from tests.util import assert_close_matrices, multivariate_rhat
-
-
-class Data(NamedTuple):
-    """Toy dataset for testing."""
-
-    X: Int32[Array, 'p n']
-    y: Float32[Array, ' n']
-    max_split: UInt32[Array, ' p']
-
-
-class TestWishart:
-    """Test the basic properties of the wishart sampler output."""
-
-    # Parameterize with (k, df) pairs
-    @pytest.fixture(params=[(1, 3), (3, 3), (3, 5), (3, 100), (100, 102)])
-    def wishart_params(self, request: FixtureRequest) -> tuple[int, int]:
-        """Provide (k, df) pairs for testing."""
-        k, df = request.param
-        return k, df
-
-    def random_pd_matrix(self, key: Key[Array, ''], k: int) -> Float[Array, '{k} {k}']:
-        """Generate a random positive definite matrix."""
-        A = random.normal(key, (k, k))
-        return A @ A.T + jnp.eye(k)
-
-    def ill_conditioned_matrix(
-        self, key: Key[Array, ''], k: int, condition_number: float = 1e6
-    ) -> Float[Array, '{k} {k}']:
-        """Generate a ill conditioned random positive semi-definite matrix."""
-        A = random.normal(key, (k, k))
-        U, _ = jnp.linalg.qr(A)
-
-        if k == 1:
-            eigs = jnp.zeros(1)
-        else:
-            smalls = jnp.geomspace(1.0, 1.0 / condition_number, num=k - 1)
-            eigs = jnp.concatenate([smalls, jnp.array([0.0])])
-        return (U * eigs) @ U.T
-
-    def test_size(self, keys: split, wishart_params: tuple[int, int]) -> None:
-        """Check that the sample generated by wishart sampler is of shape k*k."""
-        k, df = wishart_params
-        scale = self.random_pd_matrix(keys.pop(), k)
-        sample = _sample_wishart_bartlett(keys.pop(), df, scale)
-        assert sample.shape == (k, k)
-
-    def test_symmetric(self, keys: split, wishart_params: tuple[int, int]) -> None:
-        """Check that the sample generated by wishart sampler is symmetric."""
-        k, df = wishart_params
-        scale = self.random_pd_matrix(keys.pop(), k)
-        sample = _sample_wishart_bartlett(keys.pop(), df, scale)
-        assert_close_matrices(sample, sample.T, rtol=1e-6)
-
-    def test_pos_def(self, keys: split, wishart_params: tuple[int, int]) -> None:
-        """Check that the sample generated by wishart sampler is positive definite."""
-        k, df = wishart_params
-        scale = self.random_pd_matrix(keys.pop(), k)
-        sample = _sample_wishart_bartlett(keys.pop(), df, scale)
-        eigs = jnp.linalg.eigvalsh(sample)
-        assert jnp.all(eigs > 0)
-
-    def test_near_singular_scale(
-        self, keys: split, wishart_params: tuple[int, int]
-    ) -> None:
-        """Check that the wishart sampler still works with singular or near singular matrix."""
-        k, df = wishart_params
-        ill_conditioned_scale = self.ill_conditioned_matrix(keys.pop(), k)
-        sample = _sample_wishart_bartlett(keys.pop(), df, ill_conditioned_scale)
-        assert jnp.all(jnp.isfinite(sample))
-
-    def test_wishart_dist(self, keys: split, wishart_params: tuple[int, int]) -> None:
-        """Check that the sample generated by wishart sampler follows a wishart distribution."""
-        k, df = wishart_params
-        sigma = self.random_pd_matrix(keys.pop(), k)
-        scale_inv = jnp.linalg.inv(sigma)
-
-        a = random.normal(keys.pop(), (k,))
-        denominator = a.T @ sigma @ a
-
-        sampler = vmap(_sample_wishart_bartlett, in_axes=(0, None, None))
-        W = sampler(keys.pop(1000), float(df), scale_inv)
-        t = jnp.einsum('ijk,j,k->i', W, a, a) / denominator
-
-        test = ks_1samp(t, chi2(df).cdf)
-        assert test.pvalue > 0.01
-
-
-class TestPrecomputeTerms:
-    """Test _precompute_likelihood_terms_mv and _precompute_leaf_terms_mv correctness and stability."""
-
-    @pytest.fixture(params=[1, 2, 5, 10])
-    def k(self, request: FixtureRequest) -> int:
-        """Provide different ks for testing."""
-        return request.param
-
-    def random_pd_matrix(self, key: Key[Array, ''], k: int) -> Float[Array, '{k} {k}']:
-        """Generate a random positive definite matrix."""
-        A = random.normal(key, (k, k))
-        return A @ A.T + jnp.eye(k)
-
-    def test_shapes_leaf(self, keys: split, k: int) -> None:
-        """Check that shapes of outputs are correct."""
-        num_trees, tree_size = 3, 4
-        prec_trees = jnp.ones((num_trees, tree_size))
-        error_cov_inv = self.random_pd_matrix(keys.pop(), k)
-        leaf_prior_cov_inv = self.random_pd_matrix(keys.pop(), k)
-
-        result = _precompute_leaf_terms_mv(
-            keys.pop(), prec_trees, error_cov_inv, leaf_prior_cov_inv
-        )
-        assert result.mean_factor.shape == (num_trees, k, k, tree_size)
-        assert result.centered_leaves.shape == (num_trees, k, tree_size)
-
-    def test_likelihood_equiv(self, keys: split) -> None:
-        """Check that _compute_likelihood_ratio_uv and _compute_likelihood_ratio_mv agree when k = 1."""
-        inv_sigma2 = random.uniform(keys.pop(), (), minval=0.1, maxval=5.0)
-        leaf_prior_cov_inv_uv = random.uniform(keys.pop(), (), minval=0.1, maxval=5.0)
-        error_cov_inv = jnp.array([[inv_sigma2]])
-        leaf_prior_cov_inv = jnp.array([[leaf_prior_cov_inv_uv]])
-
-        precs = Counts(left=jnp.array(3.0), right=jnp.array(4.0), total=jnp.array(7.0))
-
-        total_resid = random.normal(keys.pop(), (1,))
-        left_resid = random.normal(keys.pop(), (1,))
-        right_resid = random.normal(keys.pop(), (1,))
-
-        prelkv_mv, _ = _precompute_likelihood_terms_mv(
-            error_cov_inv, leaf_prior_cov_inv, precs
-        )
-        likelihood_mv = _compute_likelihood_ratio_mv(
-            total_resid, left_resid, right_resid, prelkv_mv
-        )
-
-        prelkv_uv, prelk_uv = _precompute_likelihood_terms_uv(
-            inv_sigma2, leaf_prior_cov_inv_uv, precs
-        )
-        likelihood_uv = _compute_likelihood_ratio_uv(
-            total_resid.item(),
-            left_resid.item(),
-            right_resid.item(),
-            prelkv_uv,
-            prelk_uv,
-        )
-
-        assert_allclose(
-            prelkv_mv.log_sqrt_term, prelkv_uv.log_sqrt_term, rtol=1e-6, atol=1e-6
-        )
-        assert_allclose(likelihood_mv, likelihood_uv, rtol=1e-6, atol=1e-6)
-
-    def test_leaf_terms_equiv(self, keys: split) -> None:
-        """Check that _precompute_leaf_terms_uv and _precompute_leaf_terms_mv agree when k = 1."""
-        num_trees, tree_size = 2, 3
-        inv_sigma2 = random.uniform(keys.pop(), (), minval=0.1, maxval=5.0)
-        leaf_prior_cov_inv_uv = random.uniform(keys.pop(), (), minval=0.1, maxval=5.0)
-
-        error_cov_inv = jnp.array([[inv_sigma2]])
-        leaf_prior_cov_inv = jnp.array([[leaf_prior_cov_inv_uv]])
-        prec_trees = random.uniform(keys.pop(), (num_trees, tree_size)) * 5.0
-        z_mv = random.normal(keys.pop(), (num_trees, tree_size, 1))
-        z_uv = z_mv.squeeze(axis=-1)
-
-        result_uv = _precompute_leaf_terms_uv(
-            keys.pop(), prec_trees, inv_sigma2, leaf_prior_cov_inv_uv, z_uv
-        )
-        result_mv = _precompute_leaf_terms_mv(
-            keys.pop(), prec_trees, error_cov_inv, leaf_prior_cov_inv, z_mv
-        )
-
-        assert_allclose(
-            result_uv.mean_factor,
-            result_mv.mean_factor.squeeze((1, 2)),
-            rtol=1e-6,
-            atol=1e-6,
-        )
-        assert_allclose(
-            result_uv.centered_leaves,
-            result_mv.centered_leaves.squeeze(1),
-            rtol=1e-6,
-            atol=1e-6,
-        )
-
-
-@pytest.fixture(params=[(10, 2), (20, 5), (3, 100), (50, 50)])
-def data_shape(request: FixtureRequest) -> tuple[int, int]:
-    """Provide (n, p) pairs for testing."""
-    return request.param
-
-
-@pytest.fixture
-def data(data_shape: tuple[int, int]) -> Data:
-    """Generate a toy dataset."""
-    n, p = data_shape
-    X = jnp.arange(n * p).reshape(p, n)
-    y = jnp.linspace(-1, 1, n)
-    max_split = jnp.full(p, 5, dtype=jnp.uint32)
-    return Data(X, y, max_split)
-
-
-class TestMVBartIntegration:
-    """Test equivalence between Univariate and Multivariate (k=1) modes."""
-
-    @pytest.mark.parametrize('binary', [False, True])
-    def test_init_equivalence(self, data: Data, binary: bool) -> None:
-        """Test that init produces compatible structures for UV and MV(k=1)."""
-        X, y, max_split = data
-        p_nonterminal = jnp.array([0.9, 0.5])
-
-        if binary:
-            y = (y > 0).astype(jnp.float32)
-
-        common: dict = dict(
-            X=X,
-            max_split=max_split,
-            num_trees=10,
-            p_nonterminal=p_nonterminal,
-            resid_num_batches=None,
-            count_num_batches=None,
-        )
-
-        uv_kw = dict(y=y, offset=0.0, leaf_prior_cov_inv=1.0)
-        mv_kw = dict(
-            y=y[None, :], offset=jnp.zeros(1), leaf_prior_cov_inv=jnp.array([[1.0]])
-        )
-
-        if binary:
-            uv_kw.update(outcome_type='binary')
-            mv_kw.update(outcome_type='binary')
-        else:
-            uv_kw.update(error_cov_df=6.0, error_cov_scale=4.0)
-            mv_kw.update(error_cov_df=jnp.array(6.0), error_cov_scale=4.0 * jnp.eye(1))
-
-        bart_uv = init(**uv_kw, **common)
-        bart_mv = init(**mv_kw, **common)
-
-        assert bart_uv.resid.ndim == 1
-        assert bart_mv.resid.ndim == 2
-        assert bart_mv.resid.shape[0] == 1
-        assert bart_mv.resid.shape[1] == bart_uv.resid.shape[0]
-
-        assert jnp.ndim(bart_uv.error_cov_inv) == 0
-        assert bart_mv.error_cov_inv.shape == (1, 1)
-
-        if binary:
-            assert bart_uv.binary_y is not None
-            assert bart_mv.binary_y is not None
-            assert bart_uv.binary_y.ndim == 1
-            assert bart_mv.binary_y.ndim == 2
-            assert_array_equal(bart_uv.binary_y, bart_mv.binary_y.squeeze(0))
-            assert bart_uv.z is not None
-            assert bart_mv.z is not None
-            assert bart_uv.z.ndim == 1
-            assert bart_mv.z.ndim == 2
-            assert_array_equal(bart_uv.z, bart_mv.z.squeeze(0))
-
-        assert_array_equal(bart_uv.resid, bart_mv.resid.squeeze(0))
-        assert_array_equal(bart_uv.forest.var_tree, bart_mv.forest.var_tree)
-        assert_array_equal(bart_uv.forest.split_tree, bart_mv.forest.split_tree)
-        assert_array_equal(
-            bart_uv.forest.leaf_tree, bart_mv.forest.leaf_tree.squeeze(1)
-        )
-        assert_array_equal(bart_uv.forest.leaf_indices, bart_mv.forest.leaf_indices)
-        assert_array_equal(bart_uv.forest.p_nonterminal, bart_mv.forest.p_nonterminal)
-        assert_array_equal(bart_uv.forest.p_propose_grow, bart_mv.forest.p_propose_grow)
-        assert_array_equal(bart_uv.forest.affluence_tree, bart_mv.forest.affluence_tree)
-
-    def test_step_sigma_distribution_match(self, keys: split, data: Data) -> None:
-        """
-        Test that _step_error_cov_inv_uv and _step_error_cov_inv_mv (k = 1) sample from the same posterior.
-
-        UV: 1/sigma2 ~ Gamma(alpha_post, beta_post)
-        MV: error_cov_inv ~ Wishart(df_post, scale_post)
-        """
-        X, y, _ = data
-        resid = random.normal(keys.pop(), (y.size,))
-
-        # inverse gamma prior: alpha = df / 2, beta = scale / 2
-        df_prior = jnp.float32(20.0)
-        scale_prior = jnp.float32(10.0)
-
-        common: dict = dict(
-            X=X,
-            binary_y=None,
-            binary_indices=None,
-            error_cov_df=df_prior,
-            z=None,
-            offset=0.0,
-            prec_scale=None,
-            forest=None,
-            config=None,
-        )
-
-        st_uv = State(
-            **common, resid=resid, error_cov_scale=scale_prior, error_cov_inv=1.0
-        )
-
-        st_mv = State(
-            **common,
-            resid=resid[None, :],
-            error_cov_scale=jnp.array([[scale_prior]]),
-            error_cov_inv=jnp.eye(1),
-        )
-
-        def sample_uv(k: Key[Array, '']) -> Float32[Array, '']:
-            return _step_error_cov_inv_uv(k, st_uv).error_cov_inv
-
-        def sample_mv(k: Key[Array, '']) -> Float32[Array, '']:
-            return _step_error_cov_inv_mv(k, st_mv).error_cov_inv.reshape(())
-
-        n_samples = 10000
-        samples_uv = vmap(sample_uv)(keys.pop(n_samples))
-        samples_mv = vmap(sample_mv)(keys.pop(n_samples))
-
-        _, p_value = ks_2samp(samples_uv, samples_mv)
-
-        assert jnp.abs(jnp.mean(samples_uv) - jnp.mean(samples_mv)) < 0.01
-        assert p_value > 0.01
-
-    def test_uv_mv_k1_equivalence(self, keys: split) -> None:
-        """Test that Bart class initializes equivalent states for UV and MV (k=1)."""
-        n, p = 20, 5
-        X = random.normal(keys.pop(), (p, n))
-        y_uv = random.normal(keys.pop(), (n,))
-        y_mv = y_uv[None, :]  # shape (1, n)
-
-        common = dict(x_train=X, num_trees=10, ndpost=0, nskip=0, num_chains=1, seed=42)
-        bart_uv = Bart(y_train=y_uv, **common)
-        bart_mv = Bart(y_train=y_mv, **common)
-
-        state_uv = bart_uv._mcmc_state
-        state_mv = bart_mv._mcmc_state
-
-        # Residuals and error covariance
-        assert_allclose(state_uv.resid, state_mv.resid.squeeze(0), atol=1e-6, rtol=1e-6)
-        assert_allclose(
-            state_uv.error_cov_inv,
-            state_mv.error_cov_inv.reshape(()),
-            atol=1e-6,
-            rtol=1e-6,
-        )
-
-        # Prior parameters
-        assert_array_equal(
-            state_uv.forest.leaf_prior_cov_inv,
-            state_mv.forest.leaf_prior_cov_inv.reshape(()),
-        )
-        assert_array_equal(state_uv.error_cov_df, state_mv.error_cov_df)
-        assert_array_equal(
-            state_uv.error_cov_scale, state_mv.error_cov_scale.reshape(())
-        )
-
-        # Forest structure
-        assert_array_equal(state_uv.forest.var_tree, state_mv.forest.var_tree)
-        assert_array_equal(state_uv.forest.split_tree, state_mv.forest.split_tree)
-        assert_array_equal(
-            state_uv.forest.leaf_tree, state_mv.forest.leaf_tree.squeeze(-2)
-        )
-        assert_array_equal(state_uv.forest.leaf_indices, state_mv.forest.leaf_indices)
-
-        # Offset
-        assert_allclose(bart_uv.offset, bart_mv.offset.squeeze(0), atol=1e-6, rtol=1e-6)
-
-        # Sigest
-        assert_allclose(bart_uv.sigest, bart_mv.sigest.squeeze(0), atol=1e-6, rtol=1e-6)
-
-
-class TestMVBartSteps:
-    """Test the full MCMC step trajectory (init + multiple steps)."""
-
-    @pytest.mark.parametrize('binary', [False, True])
-    def test_step_trees_exact_match(
-        self, keys: split, data: Data, binary: bool
-    ) -> None:
-        """Test that MV tree logic is Identical to UV logic."""
-        X, y, max_split = data
-        n_trees = 100
-
-        if binary:
-            y = (y > 0).astype(jnp.float32)
-
-        params: dict = dict(
-            X=X,
-            max_split=max_split,
-            num_trees=n_trees,
-            p_nonterminal=jnp.array([0.9, 0.5]),
-            resid_num_batches=None,
-            count_num_batches=None,
-        )
-
-        uv_kw = dict(y=y, offset=0.0, leaf_prior_cov_inv=jnp.float32(n_trees))
-        mv_kw = dict(
-            y=y[None, :], offset=jnp.zeros(1), leaf_prior_cov_inv=n_trees * jnp.eye(1)
-        )
-
-        if binary:
-            uv_kw.update(outcome_type='binary')
-            mv_kw.update(outcome_type='binary')
-        else:
-            uv_kw.update(error_cov_df=4.0, error_cov_scale=2.0)
-            mv_kw.update(error_cov_df=jnp.array(4.0), error_cov_scale=2 * jnp.eye(1))
-
-        uv_state = init(**uv_kw, **params)
-        mv_state = init(**mv_kw, **params)
-
-        mv_state = replace(
-            mv_state,
-            resid=uv_state.resid[None, :],
-            error_cov_inv=jnp.array([[uv_state.error_cov_inv]]),
-            forest=replace(
-                mv_state.forest,
-                var_tree=uv_state.forest.var_tree,
-                split_tree=uv_state.forest.split_tree,
-                leaf_tree=uv_state.forest.leaf_tree[:, None, :],
-                leaf_indices=uv_state.forest.leaf_indices,
-                affluence_tree=uv_state.forest.affluence_tree,
-            ),
-        )
-
-        key = keys.pop()
-        uv_next = step_trees(key, uv_state)
-        mv_next = step_trees(random.clone(key), mv_state)
-
-        assert_close_matrices(
-            uv_next.resid, mv_next.resid.squeeze(0), atol=1e-6, rtol=1e-6
-        )
-        assert_close_matrices(
-            uv_state.forest.leaf_tree,
-            mv_state.forest.leaf_tree.squeeze(1),
-            atol=1e-6,
-            rtol=1e-6,
-        )
-
-        assert_array_equal(uv_state.forest.var_tree, mv_state.forest.var_tree)
-        assert_array_equal(uv_state.forest.split_tree, mv_state.forest.split_tree)
-        assert_array_equal(uv_state.forest.leaf_indices, mv_state.forest.leaf_indices)
-        assert_array_equal(
-            uv_state.forest.affluence_tree, mv_state.forest.affluence_tree
-        )
-
-        assert_array_equal(
-            uv_state.forest.grow_prop_count, mv_state.forest.grow_prop_count
-        )
-        assert_array_equal(
-            uv_state.forest.grow_acc_count, mv_state.forest.grow_acc_count
-        )
-        assert_array_equal(
-            uv_state.forest.prune_prop_count, mv_state.forest.prune_prop_count
-        )
-        assert_array_equal(
-            uv_state.forest.prune_acc_count, mv_state.forest.prune_acc_count
-        )
-
-    @pytest.mark.parametrize('binary', [False, True])
-    def test_mv_steps(self, keys: split, data: Data, binary: bool) -> None:
-        """Test that mv mode can run without crashing."""
-        X, y_uv, max_split = data
-        k = 3
-
-        if binary:
-            y = random.bernoulli(keys.pop(), 0.5, (k, y_uv.size)).astype(jnp.float32)
-        else:
-            y = jnp.tile(y_uv, (k, 1))
-            y = y + random.normal(keys.pop(), y.shape) * 0.1
-
-        kw: dict = dict(
-            X=X,
-            y=y,
-            offset=jnp.zeros(k),
-            max_split=max_split,
-            num_trees=5,
-            p_nonterminal=jnp.array([0.9, 0.5]),
-            leaf_prior_cov_inv=jnp.eye(k),
-            resid_num_batches=None,
-            count_num_batches=None,
-        )
-
-        if binary:
-            kw.update(outcome_type='binary')
-        else:
-            kw.update(error_cov_df=jnp.array(10.0), error_cov_scale=jnp.eye(k))
-
-        mv_state = init(**kw)
-
-        for key in keys.pop(10):
-            mv_state = step(key, mv_state)
-
-            assert jnp.all(jnp.isfinite(mv_state.resid))
-            assert jnp.all(jnp.isfinite(mv_state.forest.leaf_tree))
-
-            assert mv_state.resid.shape == (k, y.shape[1])
-
-            assert jnp.all(jnp.isfinite(mv_state.error_cov_inv))
-            assert mv_state.error_cov_inv.shape == (k, k)
-
-            if binary:
-                assert mv_state.z is not None
-                assert jnp.all(jnp.isfinite(mv_state.z))
-                assert mv_state.z.shape == (k, y.shape[1])
-
-
-class MVData(NamedTuple):
-    """Dataset for testing in `TestMVBartInterface`."""
-
-    x: Float32[Array, 'p n']
-    y: Float32[Array, 'k n']
-    w: Float32[Array, ' n'] | None = None
-    outcome_type: str | list[str] = 'continuous'
-
-
-class TestMVBartInterface:
-    """Tests for the high-level Bart class with multivariate y."""
-
-    @pytest.fixture
-    def kwargs(self) -> dict:
-        """Provide base keyword arguments for Bart initialization."""
-        return dict(num_trees=5, ndpost=10, nskip=5, num_chains=None)
-
-    @pytest.fixture(params=[None, 2])
-    def num_chains(self, request: FixtureRequest) -> int | None:
-        """Provide the number of chains to test with."""
-        return request.param
-
-    @pytest.fixture(params=[(20, 6, 2), (20, 10, 3), (50, 100, 4), (50, 50, 5)])
-    def mv_data_shape(self, request: FixtureRequest) -> tuple[int, int, int]:
-        """Provide (n, p, k) triples for testing."""
-        return request.param
-
-    @pytest.fixture
-    def mv_data(self, keys: split, mv_data_shape: tuple[int, int, int]) -> MVData:
-        """Generate a toy multivariate dataset."""
-        n, p, k = mv_data_shape
-        dgp = gen_data(
-            keys.pop(),
-            n=n,
-            p=p,
-            k=k,
-            q=2,
-            lam=0.5,
-            sigma2_lin=0.4,
-            sigma2_quad=0.5,
-            sigma2_eps=0.1,
-        )
-        return MVData(dgp.x, dgp.y)
-
-    @pytest.fixture
-    def example_data(self, keys: split) -> MVData:
-        """Return a small nonsense dataset for tests that don't need realistic data."""
-        return MVData(
-            random.normal(keys.pop(), (2, 5)),
-            random.normal(keys.pop(), (3, 5)),
-            random.normal(keys.pop(), (5,)),
-        )
-
-    @pytest.fixture(params=['continuous', 'binary', 'mixed'])
-    def mv_outcome_data(self, request: FixtureRequest, mv_data: MVData) -> MVData:
-        """Apply outcome_mode postprocessing to mv_data."""
-        outcome_mode = request.param
-        k = mv_data.y.shape[0]
-        if outcome_mode == 'binary':
-            y = (mv_data.y > 0).astype(jnp.float32)
-            outcome_type = 'binary'
-        elif outcome_mode == 'mixed':
-            y = mv_data.y.at[0].set((mv_data.y[0] > 0).astype(jnp.float32))
-            outcome_type = ['binary'] + ['continuous'] * (k - 1)
-        else:
-            y = mv_data.y
-            outcome_type = 'continuous'
-        return MVData(x=mv_data.x, y=y, outcome_type=outcome_type)
-
-    def test_initialization_and_shapes(
-        self, keys: split, mv_outcome_data: MVData, kwargs: dict, num_chains: int | None
-    ) -> None:
-        """Test that MV Bart predicts with correct shapes."""
-        k, n = mv_outcome_data.y.shape
-        n_test = 40
-        p = mv_outcome_data.x.shape[0]
-        outcome_type = mv_outcome_data.outcome_type
-
-        kwargs.update(num_chains=num_chains, outcome_type=outcome_type)
-        bart = Bart(x_train=mv_outcome_data.x, y_train=mv_outcome_data.y, **kwargs)
-
-        # test predict shape
-        x_test = random.normal(keys.pop(), (p, n_test))
-        for x, m in [('train', n), (x_test, n_test)]:
-            y_pred = bart.predict(x, kind='latent_samples')
-            assert y_pred.shape == (bart.ndpost, k, m)
-
-            y_pred = bart.predict(x, kind='mean_samples')
-            assert y_pred.shape == (bart.ndpost, k, m)
-
-            y_pred = bart.predict(x, kind='mean')
-            assert y_pred.shape == (k, m)
-
-            y_pred = bart.predict(x, kind='outcome_samples', key=keys.pop())
-            assert y_pred.shape == (bart.ndpost, k, m)
-
-            mean = bart.predict(x, kind='mean')
-            outcomes = bart.predict(x, kind='outcome_samples', key=keys.pop())
-            if outcome_type == 'binary':
-                assert jnp.all((mean >= 0) & (mean <= 1))
-                assert jnp.all((outcomes == 0) | (outcomes == 1))
-            elif isinstance(outcome_type, list):
-                assert jnp.all((mean[0] >= 0) & (mean[0] <= 1))
-                assert jnp.all((outcomes[..., 0, :] == 0) | (outcomes[..., 0, :] == 1))
-
-        # config params shape
-        assert bart.offset.shape == (k,)
-        if outcome_type == 'binary':
-            assert bart.sigest is None
-        elif isinstance(outcome_type, list):
-            sigest = bart.sigest
-            assert sigest.shape == (k,)
-            assert sigest[0] == 0.0  # binary component
-            assert jnp.all(sigest[1:] > 0)  # continuous components
-        else:
-            assert bart.sigest.shape == (k,)
-
-    @pytest.mark.parametrize('outcome_mode', ['continuous', 'mixed'])
-    def test_scalar_params(
-        self, example_data: MVData, subtests: SubTests, outcome_mode: str, kwargs: dict
-    ) -> None:
-        """Test that scalar configuration params are broadcasted."""
-        k, _ = example_data.y.shape
-        if outcome_mode == 'mixed':
-            outcome_type = ['binary'] + ['continuous'] * (k - 1)
-        else:
-            outcome_type = 'continuous'
-        kwargs.update(ndpost=0, nskip=0, outcome_type=outcome_type)
-
-        with subtests.test('offset'):
-            bart = Bart(example_data.x, example_data.y, offset=0.0, **kwargs)
-            assert bart.offset.shape == (k,)
-
-        with subtests.test('sigest'):
-            bart = Bart(example_data.x, example_data.y, sigest=1.0, **kwargs)
-            assert bart.sigest.shape == (k,)
-
-        with subtests.test('lamda'):
-            bart = Bart(example_data.x, example_data.y, lamda=1.0, **kwargs)
-            assert bart.sigest is None
-            assert bart._mcmc_state.error_cov_scale.shape == (k, k)
-
-    def test_mv_rejects_weights(self, example_data: MVData, kwargs: dict) -> None:
-        """MV + weights should raise."""
-        with pytest.raises(ValueError, match='Weights'):
-            Bart(
-                x_train=example_data.x,
-                y_train=example_data.y,
-                w=example_data.w,
-                **kwargs,
-            )
-
-    def test_get_latent_prec(
-        self, mv_outcome_data: MVData, kwargs: dict, num_chains: int | None
-    ) -> None:
-        """get_latent_prec returns correct shape, dtype, and is symmetric PD."""
-        k = mv_outcome_data.y.shape[0]
-        kwargs.update(num_chains=num_chains, outcome_type=mv_outcome_data.outcome_type)
-        bart = Bart(x_train=mv_outcome_data.x, y_train=mv_outcome_data.y, **kwargs)
-        ndpost = kwargs['ndpost']
-        nskip = kwargs['nskip']
-
-        prec = bart.get_latent_prec()
-        if num_chains is not None:
-            assert prec.shape == (num_chains, nskip + ndpost // num_chains, k, k)
-        else:
-            assert prec.shape == (nskip + ndpost, k, k)
-        assert prec.dtype == jnp.float32
-
-        # check symmetry and positive definiteness
-        assert_close_matrices(prec, prec.mT, rtol=1e-6, reduce_rank=True)
-        eigvals = jnp.linalg.eigvalsh(prec)
-        assert jnp.all(eigvals > 0)
-
-    def test_get_latent_prec_only_continuous(
-        self, mv_outcome_data: MVData, kwargs: dict, num_chains: int | None
-    ) -> None:
-        """get_latent_prec(only_continuous=True) removes binary components."""
-        k = mv_outcome_data.y.shape[0]
-        outcome_type = mv_outcome_data.outcome_type
-        kwargs.update(num_chains=num_chains, outcome_type=outcome_type)
-        bart = Bart(x_train=mv_outcome_data.x, y_train=mv_outcome_data.y, **kwargs)
-        ndpost = kwargs['ndpost']
-        nskip = kwargs['nskip']
-
-        if outcome_type == 'binary':
-            with pytest.raises(ValueError, match='only binary'):
-                bart.get_latent_prec(only_continuous=True)
-            return
-
-        prec = bart.get_latent_prec(only_continuous=True)
-        if isinstance(outcome_type, list):
-            kb = sum(1 for t in outcome_type if t == 'binary')
-            kc = k - kb
-        else:
-            kc = k
-        if num_chains is not None:
-            assert prec.shape == (num_chains, nskip + ndpost // num_chains, kc, kc)
-        else:
-            assert prec.shape == (nskip + ndpost, kc, kc)
-
-    def test_get_error_sdev_shape(
-        self, mv_outcome_data: MVData, kwargs: dict, num_chains: int | None
-    ) -> None:
-        """get_error_sdev returns correct shape and NaN pattern."""
-        k = mv_outcome_data.y.shape[0]
-        outcome_type = mv_outcome_data.outcome_type
-        kwargs.update(num_chains=num_chains, outcome_type=outcome_type)
-        bart = Bart(x_train=mv_outcome_data.x, y_train=mv_outcome_data.y, **kwargs)
-        ndpost = kwargs['ndpost']
-
-        with debug_nans(False):
-            sdev = bart.get_error_sdev()
-            assert sdev.shape == (ndpost, k)
-
-            if outcome_type == 'binary':
-                assert jnp.all(jnp.isnan(sdev))
-            elif isinstance(outcome_type, list):
-                binary_mask = jnp.array([t == 'binary' for t in outcome_type])
-                assert jnp.all(jnp.isnan(sdev[:, binary_mask]))
-                assert jnp.all(jnp.isfinite(sdev[:, ~binary_mask]))
-            else:
-                assert jnp.all(jnp.isfinite(sdev))
-                assert jnp.all(sdev > 0)
-
-    def test_get_error_sdev_mean(
-        self, mv_outcome_data: MVData, kwargs: dict, num_chains: int | None
-    ) -> None:
-        """get_error_sdev(mean=True) returns correct shape and NaN pattern."""
-        k = mv_outcome_data.y.shape[0]
-        outcome_type = mv_outcome_data.outcome_type
-        kwargs.update(num_chains=num_chains, outcome_type=outcome_type)
-        bart = Bart(x_train=mv_outcome_data.x, y_train=mv_outcome_data.y, **kwargs)
-
-        with debug_nans(False):
-            sdev_mean = bart.get_error_sdev(mean=True)
-            assert sdev_mean.shape == (k,)
-
-            if outcome_type == 'binary':
-                assert jnp.all(jnp.isnan(sdev_mean))
-            elif isinstance(outcome_type, list):
-                binary_mask = jnp.array([t == 'binary' for t in outcome_type])
-                assert jnp.all(jnp.isnan(sdev_mean[binary_mask]))
-                assert jnp.all(jnp.isfinite(sdev_mean[~binary_mask]))
-            else:
-                assert jnp.all(jnp.isfinite(sdev_mean))
-                assert jnp.all(sdev_mean > 0)
-
-    def test_get_error_sdev_values(
-        self, mv_data: MVData, kwargs: dict, num_chains: int | None
-    ) -> None:
-        """get_error_sdev matches manual computation from precision matrices."""
-        kwargs.update(num_chains=num_chains)
-        bart = Bart(x_train=mv_data.x, y_train=mv_data.y, **kwargs)
-        nskip = kwargs['nskip']
-        sdev = bart.get_error_sdev()
-
-        # manual: invert each precision matrix, take sqrt of diagonal
-        prec = bart.get_latent_prec()
-        if num_chains is not None:
-            prec_post = prec[:, nskip:]  # skip burnin per chain
-            prec_post = prec_post.reshape(-1, *prec_post.shape[2:])  # flatten chains
-        else:
-            prec_post = prec[nskip:]  # skip burnin
-        cov = jnp.linalg.inv(prec_post)
-        expected = jnp.sqrt(jnp.diagonal(cov, axis1=-2, axis2=-1))
-        assert_close_matrices(sdev, expected, rtol=1e-5)
-
-    def test_mixed_rejects_weights(self, example_data: MVData, kwargs: dict) -> None:
-        """Mixed outcome_type + weights should raise."""
-        k, _ = example_data.y.shape
-        kwargs.update(outcome_type=['binary'] + ['continuous'] * (k - 1))
-        with pytest.raises(ValueError, match='univariate continuous'):
-            Bart(
-                x_train=example_data.x,
-                y_train=example_data.y,
-                w=example_data.w,
-                **kwargs,
-            )
-
-    def test_outcome_type_length_mismatch(
-        self, example_data: MVData, kwargs: dict
-    ) -> None:
-        """Sequence outcome_type with wrong length should raise."""
-        kwargs.update(outcome_type=['continuous', 'continuous'])
-        with pytest.raises(ValueError, match='length'):
-            Bart(x_train=example_data.x, y_train=example_data.y, **kwargs)
-
-    def test_sequence_outcome_type_requires_2d(self, keys: split, kwargs: dict) -> None:
-        """Sequence outcome_type with 1D y should raise."""
-        x = random.normal(keys.pop(), (2, 5))
-        y = random.normal(keys.pop(), (5,))
-        kwargs.update(outcome_type=['continuous'])
-        with pytest.raises(ValueError, match=r'y_train\.shape=\(1, n\)'):
-            Bart(x_train=x, y_train=y, **kwargs)
-
-    def test_mvbart_convergence(self, mv_data: MVData, keys: split) -> None:
-        """Test that MV Bart chains converge using R-hat."""
-        _, n_train = mv_data.x.shape
-        k_dim = mv_data.y.shape[0]
-
-        num_chains = 4
-        ndpost = 2000
-        nsamples_per_chain = ndpost // num_chains
-        nskip = 4000
-        keepevery = 5
-        num_trees = 100
-
-        bart = Bart(
-            x_train=mv_data.x,
-            y_train=mv_data.y,
-            num_trees=num_trees,
-            ndpost=ndpost,
-            nskip=nskip,
-            keepevery=keepevery,
-            num_chains=num_chains,
-            seed=keys.pop(),
-        )
-
-        # Check yhat convergence
-        yhat_train = bart.predict('train', kind='latent_samples')
-        yhat_train = yhat_train.reshape(num_chains, nsamples_per_chain, k_dim * n_train)
-        rhat_yhat_train = multivariate_rhat(yhat_train)
-        assert rhat_yhat_train < 1.6
-        print(f'{rhat_yhat_train.item()=}')
-
-        # Check covariance matrix convergence
-        prec_trace = bart._main_trace.error_cov_inv
-        if prec_trace.ndim == 3:
-            prec_trace = prec_trace.reshape(
-                num_chains, nsamples_per_chain, k_dim, k_dim
-            )
-        prec_flat = prec_trace.reshape(num_chains, nsamples_per_chain, -1)
-        assert jnp.all(jnp.std(prec_flat, axis=1) > 1e-8), 'Sigma is not updating!'
-        rhat_prec = multivariate_rhat(prec_flat)
-        assert rhat_prec < 1.1
-        print(f'{rhat_prec.item()=}')
